+++ packages/app/src/features/actions/settings-dialog/components/SettingsDialog.tsx
@@ -1,6 +1,6 @@
 import { MultiPanelDialog } from '@/features/dialogs/common/components/MultiPanelDialog'
 import { Dialog, DialogContent, DialogTitle, DialogTrigger } from '@/ui/atoms/dialog/Dialog'
-import { IconButton } from '@/ui/atoms/new/icon-button/IconButton'
+import { IconButton } from '@/ui/atoms/icon-button/IconButton'
 import { Switch } from '@/ui/atoms/new/switch/Switch'
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/ui/atoms/tooltip/Tooltip'
 import { testIds } from '@/ui/utils/testIds'
