+++ packages/app/src/ui/atoms/dialog/Dialog.tsx
@@ -5,7 +5,7 @@ import * as DialogPrimitive from '@radix-ui/react-dialog'
 import { VariantProps, cva } from 'class-variance-authority'
 import { X } from 'lucide-react'
 import * as React from 'react'
-import { IconButton } from '../new/icon-button/IconButton'
+import { IconButton } from '../icon-button/IconButton'
 
 const Dialog = DialogPrimitive.Root
